refact: use Robolectric at :feature:gallery
