--- feature/gallery/src/androidTest/AndroidManifest.xml
@@ -1,21 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<!--
-  ~ Copyright © 2023–2024 Orcinus
-  ~
-  ~ This program is free software: you can redistribute it and/or modify it under the terms of the
-  ~ GNU General Public License as published by the Free Software Foundation, either version 3 of the
-  ~ License, or (at your option) any later version.
-  ~
-  ~ This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
-  ~ without even the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See
-  ~ the GNU General Public License for more details.
-  ~
-  ~ You should have received a copy of the GNU General Public License along with this program. If
-  ~ not, see https://www.gnu.org/licenses.
-  -->
-
-<manifest xmlns:android="http://schemas.android.com/apk/res/android">
-    <application>
-        <activity android:name="br.com.orcinus.orca.feature.gallery.test.activity.GalleryActivity" />
-    </application>
-</manifest>
